9-11

+++ src/events/event.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 import { CurrentUser } from 'src/auth/current-user.decorator';
@@ -16,10 +16,10 @@ import { UpdateEventDTO } from './input/update-event.dto';
 export class EventController {
   private readonly logger = new Logger(EventController.name);
   constructor(
-    @InjectRepository(EventEnt)
-    private readonly repository: Repository<EventEnt>,
-    @InjectRepository(Attendee)
-    private readonly attendeerepository: Repository<Attendee>,
+    // @InjectRepository(EventEnt)
+    // private readonly repository: Repository<EventEnt>,
+    // @InjectRepository(Attendee)
+    // private readonly attendeerepository: Repository<Attendee>,
     private readonly eventService:EventService
   ) {}
 
@@ -38,38 +38,38 @@ export class EventController {
     return events
   }
 
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [{
-        id: MoreThan(3),
-        when: MoreThan(new Date('2021-02-12T13:00:00'))
-      },{
-        description: Like('%meet%')
-      }],
-      take: 2,
-      order: {
-        id: 'DESC'
-      }
-    });
-  } 
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repository.find({
+  //     select: ['id', 'when'],
+  //     where: [{
+  //       id: MoreThan(3),
+  //       when: MoreThan(new Date('2021-02-12T13:00:00'))
+  //     },{
+  //       description: Like('%meet%')
+  //     }],
+  //     take: 2,
+  //     order: {
+  //       id: 'DESC'
+  //     }
+  //   });
+  // } 
 
-  @Get('/practice2')
-  async practice2() {
-    const event = await this.repository.findOne(1,{
-      relations: ['attendees']
-    });
+  // @Get('/practice2')
+  // async practice2() {
+  //   const event = await this.repository.findOne(1,{
+  //     relations: ['attendees']
+  //   });
 
-    const attendee = new Attendee()
-    attendee.name = "cascade";
-    //attendee.event = event;
+  //   const attendee = new Attendee()
+  //   attendee.name = "cascade";
+  //   //attendee.event = event;
 
-    event.attendees.push(attendee);
-    await this.repository.save(event);
+  //   event.attendees.push(attendee);
+  //   await this.repository.save(event);
 
-    return event;
-  } 
+  //   return event;
+  // } 
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -96,32 +96,39 @@ export class EventController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDTO) {
-    const event = await this.repository.findOne(id)
+  @UseGuards(AuthGuardJwt)
+  async update(@Param('id') id, @Body() input: UpdateEventDTO, @CurrentUser() user:User) {
+    const event = await this.eventService.getEvent(id)
     if(!event){
       throw new NotFoundException();
     }
 
+    if(event.organizerId !== user.id){
+      throw new ForbiddenException(null, "You not authorized to update");
+    }
 
-    return await this.repository.save( {
-      ...event,
-      ...input,
-      when: input.when ?  new Date(input.when) : event.when,
-    });
+
+    return await this.eventService.updateEvent(input,event);
 
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventService.deleteEvent(id);
-
-    if(result.affected !== 1){
+  async remove(@Param('id') id, @CurrentUser() user:User) {
+    const event = await this.eventService.getEvent(id)
+    if(!event){
       throw new NotFoundException();
     }
 
-    
-    // return result;
+    if(event.organizerId !== user.id){
+      throw new ForbiddenException(null, "You not authorized to delete");
+    }
+
+
+     await this.eventService.deleteEvent(id);
+
+
   }
 
 }
